+++ packages/categories/i18n/de.js
@@ -1,5 +1,5 @@
 
 i18n.map('de', {
-	label_cats_place: "Typ",
+	label_cats_place: "Kategorien",
 	label_cats_user: "Typologie",
 });