+++ packages/categories/i18n/en.js
@@ -1,5 +1,5 @@
 
 i18n.map('en', {
-	label_cats_place: "Type",
+	label_cats_place: "Categories",
 	label_cats_user: "Tipology",
 });