+++ packages/categories/i18n/es.js
@@ -1,5 +1,5 @@
 
 i18n.map('es', {
-	label_cats_place: "Tipo",
+	label_cats_place: "Categorías",
 	label_cats_user: "Tipologia"	
 });