+++ packages/categories/i18n/fr.js
@@ -1,5 +1,5 @@
 
 i18n.map('fr', {
-	label_cats_place: "Type",
+	label_cats_place: "Catégories",
 	label_cats_user: "Typologie",	
 });