+++ packages/categories/i18n/it.js
@@ -1,5 +1,5 @@
 
 i18n.map('it', {
-	label_cats_place: "Tipo",
+	label_cats_place: "Categorie",
 	label_cats_user: "Tipologia",
 });