+++ packages/categories/i18n/sv.js
@@ -1,5 +1,5 @@
 
 i18n.map('sv', {
-	label_cats_place: "Typ",
+	label_cats_place: "Kategorier",
 	label_cats_user: "Typologi",
 });