+++ packages/photos/i18n/de.js
@@ -1,5 +1,7 @@
 
 i18n.map('de', {
+	title_photos: "Fotos",
+
 	btn_avatar: "Profilfoto",
 	upload_error_imageNotValid: "Nur JPG/PNG-Dateien",
 });