+++ packages/photos/i18n/es.js
@@ -1,5 +1,7 @@
 
 i18n.map('es', {
+	title_photos: "Fotografías",
+
 	btn_avatar: "Foto del Perfil",
 	upload_error_imageNotValid: "solo imagen JPG/PNG",
 });