+++ packages/photos/i18n/fr.js
@@ -1,5 +1,7 @@
 
 i18n.map('fr', {
+	title_photos: "Photos",
+	
 	btn_avatar: "Photo profile",
 	upload_error_imageNotValid: "Seulement images JPG/PNG",
 });