+++ packages/photos/i18n/it.js
@@ -1,5 +1,7 @@
 
 i18n.map('it', {
+	title_photos: "Fotografie",
+	
 	btn_avatar: "Foto profilo",
 	upload_error_imageNotValid: "Solo immagini JPG/PNG",
 });