+++ packages/photos/i18n/sv.js
@@ -1,5 +1,7 @@
 
 i18n.map('sv', {
+	title_photos: "Fotografier",
+
 	btn_avatar: "Profilbild",
 	upload_error_imageNotValid: "Endast JPG/PNG-bilder",
 });