+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 - new method K.Util.geo.parseLocString to parse DMS locations
 - plugin import remove button in places list
 - plugin photos use npm imagemagick
+- plugin photos split templates in edit panel and tab place
+- fix place edit categories
 
 ## v1.5.4
 - new plugin photos